chore(desktop): split Smart Turn cache out of models.rs
The file-size ratchet caps src-tauri Rust at 1000 lines. models.rs was
994 on main and 1137 after the flag-gated cache; the download lives in a
sibling module so the STT/TTS manager stays under the ceiling.
